+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Features
 + Rearranged some context menu buttons to prevent accidental pings etc.
++ Added new buttons for left menu and plugins in settings to indicate the scroll area (work in newer versions of OpenStarbound)
 
 ## API
 + Introduced internal forceAvatar parameter